refactor: Move ManagementController into its own file
